third_party/WebKit/LayoutTests/flag-specific/enable-blink-features=LayoutNG/external/wpt/css/css-grid/alignment/grid-column-axis-self-baseline-synthesized-001-expected.txt

This is a testharness.js-based test.
PASS .grid 1
PASS .grid 2
FAIL .grid 3 assert_equals: 
<div class="grid alignItemsBaseline verticalLR">
  <div class="firstRowFirstColumn horizontalTB" data-offset-x="35" data-offset-y="0" data-expected-width="200" data-expected-height="100">ÉÉ É ÉÉÉ É ÉÉ É</div>
  <div class="firstRowSecondColumn bigFont paddingLeft" data-offset-x="0" data-offset-y="100" data-expected-width="125" data-expected-height="200">É É ÉÉ</div>
  <div class="autoRowAutoColumnSpanning2 width25"></div>
</div>
offsetLeft expected 35 but got 65
PASS .grid 4
Harness: the test ran to completion.
